code cleaning (mainly javadocs, unnecessary assignments or castings, use of StringBuilder instead of StringBuffer)

+++ hapi-base/src/main/java/ca/uhn/hl7v2/model/primitive/IS.java
@@ -75,7 +75,7 @@ public abstract class IS extends AbstractPrimitive {
      */
     public IS(Message theMessage, Integer theTable) {
         super(theMessage);
-        myTable = theTable.intValue();
+        myTable = theTable;
     }
     
     /**
